+++ CHANGELOG.md
@@ -11,8 +11,9 @@
 
 ## Bug fixes
 
-- Safety checker handles the `#copy` operator
-  ([PR #312](https://github.com/jasmin-lang/jasmin/pull/312);
+- Safety checker handles the `#copy` and `#randombytes` operators
+  ([PR #312](https://github.com/jasmin-lang/jasmin/pull/312),
+  [PR #317](https://github.com/jasmin-lang/jasmin/pull/317);
   fixes [#308](https://github.com/jasmin-lang/jasmin/issues/308)).
 
 - Register allocation takes into account conflicts between flag registers